refactor(backend): unifica portfolioHistoryRepository e portfolioSnapshotRepository
Merge dos dois repositories em portfolioRepository — leitura do histórico
e escrita de snapshots eram dois lados da mesma responsabilidade.
